+++ packages/shared/src/react/hooks/useOrganization.tsx
@@ -1,10 +1,12 @@
 /* eslint-disable jsdoc/require-description-complete-sentence */
 import type {
   ClerkPaginatedResponse,
+  CommerceSubscriptionResource,
   GetDomainsParams,
   GetInvitationsParams,
   GetMembershipRequestParams,
   GetMembersParams,
+  GetSubscriptionsParams,
   OrganizationDomainResource,
   OrganizationInvitationResource,
   OrganizationMembershipRequestResource,
@@ -64,6 +66,16 @@ export type UseOrganizationParams = {
    * </ul>
    */
   invitations?: true | PaginatedHookConfig<GetInvitationsParams>;
+  /**
+   * @experimental This is an experimental API for the Billing feature that is available under a public beta, and the API is subject to change.
+   * If set to `true`, all default properties will be used.<br />
+   * Otherwise, accepts an object with the following optional properties:
+   * <ul>
+   *  <li>`orgId`: A string that filters the subscriptions by the provided organization ID.</li>
+   *  <li>Any of the properties described in [Shared properties](#shared-properties).</li>
+   * </ul>
+   */
+  subscriptions?: true | PaginatedHookConfig<GetSubscriptionsParams>;
 };
 
 /**
@@ -99,6 +111,11 @@ export type UseOrganizationReturn<T extends UseOrganizationParams> =
        * Includes a paginated list of the organization's invitations.
        */
       invitations: PaginatedResourcesWithDefault<OrganizationInvitationResource>;
+      /**
+       * @experimental This is an experimental API for the Billing feature that is available under a public beta, and the API is subject to change.
+       * Includes a paginated list of the organization's subscriptions.
+       */
+      subscriptions: PaginatedResourcesWithDefault<CommerceSubscriptionResource>;
     }
   | {
       isLoaded: true;
@@ -108,6 +125,7 @@ export type UseOrganizationReturn<T extends UseOrganizationParams> =
       membershipRequests: PaginatedResourcesWithDefault<OrganizationMembershipRequestResource>;
       memberships: PaginatedResourcesWithDefault<OrganizationMembershipResource>;
       invitations: PaginatedResourcesWithDefault<OrganizationInvitationResource>;
+      subscriptions: PaginatedResourcesWithDefault<CommerceSubscriptionResource>;
     }
   | {
       isLoaded: boolean;
@@ -128,6 +146,10 @@ export type UseOrganizationReturn<T extends UseOrganizationParams> =
       invitations: PaginatedResources<
         OrganizationInvitationResource,
         T['invitations'] extends { infinite: true } ? true : false
+      > | null;
+      subscriptions: PaginatedResources<
+        CommerceSubscriptionResource,
+        T['subscriptions'] extends { infinite: true } ? true : false
       > | null;
     };
 
@@ -278,6 +300,7 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
     membershipRequests: membershipRequestsListParams,
     memberships: membersListParams,
     invitations: invitationsListParams,
+    subscriptions: subscriptionsListParams,
   } = params || {};
 
   useAssertWrappedByClerkProvider('useOrganization');
@@ -314,6 +337,13 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
     initialPage: 1,
     pageSize: 10,
     status: ['pending'],
+    keepPreviousData: false,
+    infinite: false,
+  });
+
+  const subscriptionsSafeValues = useWithSafeValues(subscriptionsListParams, {
+    initialPage: 1,
+    pageSize: 10,
     keepPreviousData: false,
     infinite: false,
   });
@@ -357,6 +387,15 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
           initialPage: invitationsSafeValues.initialPage,
           pageSize: invitationsSafeValues.pageSize,
           status: invitationsSafeValues.status,
+        };
+
+  const subscriptionsParams =
+    typeof subscriptionsListParams === 'undefined'
+      ? undefined
+      : {
+          initialPage: subscriptionsSafeValues.initialPage,
+          pageSize: subscriptionsSafeValues.pageSize,
+          orgId: organization?.id,
         };
 
   const domains = usePagesOrInfinite<GetDomainsParams, ClerkPaginatedResponse<OrganizationDomainResource>>(
@@ -424,6 +463,25 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
     },
   );
 
+  const subscriptions = usePagesOrInfinite<
+    GetSubscriptionsParams,
+    ClerkPaginatedResponse<CommerceSubscriptionResource>
+  >(
+    {
+      ...subscriptionsParams,
+    },
+    organization?.getSubscriptions,
+    {
+      keepPreviousData: subscriptionsSafeValues.keepPreviousData,
+      infinite: subscriptionsSafeValues.infinite,
+      enabled: !!subscriptionsParams,
+    },
+    {
+      type: 'subscriptions',
+      organizationId: organization?.id,
+    },
+  );
+
   if (organization === undefined) {
     return {
       isLoaded: false,
@@ -433,6 +491,7 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
       membershipRequests: undefinedPaginatedResource,
       memberships: undefinedPaginatedResource,
       invitations: undefinedPaginatedResource,
+      subscriptions: undefinedPaginatedResource,
     };
   }
 
@@ -445,6 +504,7 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
       membershipRequests: null,
       memberships: null,
       invitations: null,
+      subscriptions: null,
     };
   }
 
@@ -458,6 +518,7 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
       membershipRequests: undefinedPaginatedResource,
       memberships: undefinedPaginatedResource,
       invitations: undefinedPaginatedResource,
+      subscriptions: undefinedPaginatedResource,
     };
   }
 
@@ -470,5 +531,6 @@ export function useOrganization<T extends UseOrganizationParams>(params?: T): Us
     membershipRequests,
     memberships,
     invitations,
+    subscriptions,
   };
 }